Implemented column reordering.
